refactor drivers util methods
In order to install device drivers, the following workflow is followed-
- Make the device driver available within the UVM via SCSI or VSMB.
- Use `install-drivers` for LCOW and `pnputils` for WCOW to install the driver

Presently, everything was clustered together in `internal/devices` package. The utility methods to perform step 2 above are not tied to devices per se and belong to the `drivers`. Since we can easily re-use these methods, we are refactoring them into `internal/controller/drivers` where they exist as utility methods and are used in both new and old shims.
