+++ internal/devices/drivers.go
@@ -5,12 +5,9 @@ package devices
 
 import (
 	"context"
-	"errors"
 	"fmt"
 
-	"github.com/Microsoft/go-winio/pkg/guid"
-	"github.com/Microsoft/hcsshim/internal/cmd"
-	"github.com/Microsoft/hcsshim/internal/guestpath"
+	"github.com/Microsoft/hcsshim/internal/controller/drivers"
 	"github.com/Microsoft/hcsshim/internal/log"
 	"github.com/Microsoft/hcsshim/internal/resources"
 	"github.com/Microsoft/hcsshim/internal/uvm"
@@ -51,7 +48,7 @@ func InstallDrivers(ctx context.Context, vm *uvm.UtilityVM, share string) (close
 		}
 		// attempt to install even if the driver has already been installed before so we
 		// can guarantee the device is ready for use afterwards
-		return closer, execPnPInstallDriver(ctx, vm, uvmPath)
+		return closer, drivers.ExecPnPInstallDriver(ctx, vm, uvmPath)
 	}
 
 	// first mount driver as scsi in standard mount location
@@ -69,63 +66,6 @@ func InstallDrivers(ctx context.Context, vm *uvm.UtilityVM, share string) (close
 	closer = mount
 	uvmPathForShare := mount.GuestPath()
 
-	// construct path that the drivers will be remounted as read/write in the UVM
-
-	// 914aadc8-f700-4365-8016-ddad0a9d406d. Random GUID chosen for namespace.
-	ns := guid.GUID{Data1: 0x914aadc8, Data2: 0xf700, Data3: 0x4365, Data4: [8]byte{0x80, 0x16, 0xdd, 0xad, 0x0a, 0x9d, 0x40, 0x6d}}
-	driverGUID, err := guid.NewV5(ns, []byte(share))
-	if err != nil {
-		return closer, fmt.Errorf("failed to create a guid path for driver %+v: %w", share, err)
-	}
-	uvmReadWritePath := fmt.Sprintf(guestpath.LCOWGlobalDriverPrefixFmt, driverGUID.String())
-
 	// install drivers using gcs tool `install-drivers`
-	return closer, execGCSInstallDriver(ctx, vm, uvmPathForShare, uvmReadWritePath)
-}
-
-func execGCSInstallDriver(ctx context.Context, vm *uvm.UtilityVM, driverDir string, driverReadWriteDir string) error {
-	p, l, err := cmd.CreateNamedPipeListener()
-	if err != nil {
-		return err
-	}
-	defer l.Close()
-
-	var stderrOutput string
-	errChan := make(chan error)
-
-	go readAllPipeOutput(l, errChan, &stderrOutput)
-
-	args := []string{
-		"/bin/install-drivers",
-		driverReadWriteDir,
-		driverDir,
-	}
-	req := &cmd.CmdProcessRequest{
-		Args:   args,
-		Stderr: p,
-	}
-
-	// A call to `ExecInUvm` may fail in the following ways:
-	// - The process runs and exits with a non-zero exit code. In this case we need to wait on the output
-	//   from stderr so we can log it for debugging.
-	// - There's an error trying to run the process. No need to wait for stderr logs.
-	// - There's an error copying IO. No need to wait for stderr logs.
-	//
-	// Since we cannot distinguish between the cases above, we should always wait to read the stderr output.
-	exitCode, execErr := vm.ExecInUVM(ctx, req)
-
-	// wait to finish parsing stdout results
-	select {
-	case err := <-errChan:
-		if err != nil && !errors.Is(err, noExecOutputErr) {
-			return fmt.Errorf("failed to get stderr output from command %s: %w", driverDir, err)
-		}
-	case <-ctx.Done():
-		return fmt.Errorf("timed out waiting for the console output from installing driver %s: %w", driverDir, ctx.Err())
-	}
-
-	if execErr != nil {
-		return fmt.Errorf("%w: failed to install driver %s in uvm with exit code %d: %v", execErr, driverDir, exitCode, stderrOutput)
-	}
-	return nil
+	return closer, drivers.ExecGCSInstallDriver(ctx, vm, share, uvmPathForShare)
 }
