+++ internal/vm/guestmanager/manager.go
@@ -34,8 +34,8 @@ type Manager interface {
 	DumpStacks(ctx context.Context) (string, error)
 	// DeleteContainerState removes persisted state for the container identified by `cid` from the guest.
 	DeleteContainerState(ctx context.Context, cid string) error
-	// ExecIntoUVM executes commands specified in the requests in the utility VM.
-	ExecIntoUVM(ctx context.Context, request *cmd.CmdProcessRequest) (int, error)
+	// ExecInUVM executes commands specified in the requests in the utility VM.
+	ExecInUVM(ctx context.Context, request *cmd.CmdProcessRequest) (int, error)
 }
 
 var _ Manager = (*Guest)(nil)
@@ -85,7 +85,7 @@ func (gm *Guest) DeleteContainerState(ctx context.Context, cid string) error {
 	return nil
 }
 
-// ExecIntoUVM executes commands specified in the requests in the utility VM.
-func (gm *Guest) ExecIntoUVM(ctx context.Context, request *cmd.CmdProcessRequest) (int, error) {
+// ExecInUVM executes commands specified in the requests in the utility VM.
+func (gm *Guest) ExecInUVM(ctx context.Context, request *cmd.CmdProcessRequest) (int, error) {
 	return cmd.ExecInUvm(ctx, gm.gc, request)
 }
